requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
python-dotenv
requests